Fixed bug missing 0 frequency, improved peak find
There was a long-standing bug wherein the 0 frequency was not being included in the calculation of the series of frequencies using the Nyquist frequency.  The peak find code was inefficient and was cleaned up.  In addition, the plotting range now looks for the first and last peak to set the range